refactor: 주문 상세내역 OrderHistory컴포넌트로 분리
